KGC-73 Test and debug model and data

+++ Scripts/DataManager/GraphConstructor/TagDepTokenGraphConstructor.py
@@ -84,7 +84,7 @@ class TagDepTokenGraphConstructor(GraphConstructor):
         return -1 # means not found
     def __build_initial_dependency_vectors(self , dep_length : int):
         # return torch.zeros((dep_length, self.nlp.vocab.vectors_length), dtype=torch.float32)
-        return torch.nn.functional.one_hot(torch.arange(0, dep_length), num_classes=-1)
+        return torch.nn.functional.one_hot(torch.arange(0, dep_length), num_classes=-1).to(torch.float32)
     def __find_tag_index(self , tag : str):
         for tag_idx in range(len(self.tags)):
             if self.tags[tag_idx] == tag:
@@ -93,7 +93,7 @@ class TagDepTokenGraphConstructor(GraphConstructor):
     
     def __build_initial_tag_vectors(self , tags_length : int):
         # return torch.zeros((tags_length, self.nlp.vocab.vectors_length), dtype=torch.float32)
-        return torch.nn.functional.one_hot(torch.arange(0, tags_length), num_classes=-1)
+        return torch.nn.functional.one_hot(torch.arange(0, tags_length), num_classes=-1).to(torch.float32)
     
     def __build_initial_general_vector(self):
         return torch.zeros((1 , self.nlp.vocab.vectors_length), dtype=torch.float32)
